code refactoré et seeds ajoutées

+++ projet-code/back/src/main/java/com/simplon/nooky/project/controllers/AddressController.java
@@ -36,7 +36,7 @@ public class AddressController {
 	}
 	
 	@GetMapping("/{id}")
-	public Optional<Address> getAddressesById(@PathVariable @NonNull Long id) {
+	public Optional<Address> getAddressesById(@PathVariable @NonNull int id) {
 		return service.getAddressesById(id);
 	}
 }
